+++ src/helpers.js
@@ -1,8 +1,11 @@
 const vscode = require("vscode");
+const { exec } = require('child_process');
+const { promisify } = require('util');
 
 const { EXTENSION_ID, EXTENSION_NAME } = require("./constants");
 
 function determineScope(document) {
+    // TODO vscode.workspace.getWorkspaceFolder ?
     return vscode.workspace.workspaceFolders.find(scope => document.uri.fsPath.includes(scope.uri.fsPath));
 }
 
@@ -121,4 +124,6 @@ function wrapExec(exec, platform = process.platform) {
     };
 }
 
-module.exports = { determineScope, getConfiguration, commandGuard, definitionToSymbolInformation, wrapExec };
+const tryExec = wrapExec(promisify(exec));
+
+module.exports = { determineScope, getConfiguration, commandGuard, definitionToSymbolInformation, wrapExec, tryExec };